requirements.txt
absl-py==0.1.10
alabaster==0.7.10
alembic==0.9.7
anaconda-client==1.6.3
anaconda-navigator==1.6.2
anaconda-project==0.6.0
aniso8601==2.0.1
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.22.0
astroid==1.4.9
astropy==1.3.2
Babel==2.4.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
backports.weakref==1.0.post1
banal==0.3.3
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
Bottleneck==1.2.1
bs4==0.0.1
cachetools==2.0.1
cdecimal==2.3
certifi==2017.11.5
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.9
configparser==3.5.0
contextlib2==0.5.5
cryptography==1.8.1
cssselect==1.0.3
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.3
dataset==1.0.5
datashape==0.5.4
decorator==4.0.11
distributed==1.16.3
docutils==0.13.1
entrypoints==0.2.2
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12.2
Flask-Cors==3.0.2
Flask-RESTful==0.3.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
get==0.0.39
gevent==1.2.1
google-api-core==0.1.3
google-api-python-client==1.6.4
google-auth==1.2.1
google-cloud-vision==0.29.0
googleapis-common-protos==1.5.3
greenlet==0.4.12
grin==1.2.1
grpcio==1.8.2
h5py==2.7.0
HeapDict==1.0.0
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
ipaddress==1.0.18
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.1.4
krakenex==2.0.0
lazy-object-proxy==1.2.2
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.0.2
mistune==0.7.4
mock==2.0.0
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
mysql-connector-python==8.0.5
mysqlclient==1.3.12
navigator-updater==0.1.0
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.3
normality==0.5.11
nose==1.3.7
notebook==5.0.0
numba==0.33.0+0.ge79330a.dirty
numexpr==2.6.2
numpy==1.12.1
numpydoc==0.6.0
oauth2client==4.1.2
odo==0.5.0
olefile==0.44
opencv-python==3.4.0.12
openpyxl==2.4.7
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
partd==0.3.8
pathlib2==2.2.1
patsy==0.4.1
pbr==3.1.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
ply==3.10
post==0.0.26
praw==5.3.0
prawcore==0.13.0
prompt-toolkit==1.0.14
protobuf==3.5.1
psutil==5.2.2
ptyprocess==0.5.1
public==0.0.65
py==1.4.33
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.7
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
pymongo==3.6.0
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pytest==3.0.7
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
query-string==0.0.28
rake-nltk==1.0.2
readability-lxml==0.6.2
request==0.0.26
requests==2.18.4
rope==0.9.4
rsa==3.4.2
scandir==1.5
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
statsmodels==0.8.0
subprocess32==3.2.7
sympy==1.0
tables==3.3.0
tblib==1.3.2
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.6
testpath==0.3
toolz==0.8.2
tornado==4.5.1
traitlets==4.3.2
unicodecsv==0.14.1
Unidecode==0.4.21
update-checker==0.16
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==2.0.0
wrapt==1.10.10
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
zict==0.1.2
